Experimenting with client-side WASM calculation
This was a try to move the calculation to the client side using WASM. Unfortunately, this does not work as
I expected: WASM cannot execute goroutines concurrently, as the WASM / JS event loop is running in one
thread only.

So the WASM processing experiment failed in my opinion... I let this branch here for documentation.